[Port] Configure baud rate

+++ src/main/java/org/llschall/ardwloop/IArdwConfig.java
@@ -3,13 +3,13 @@ package org.llschall.ardwloop;
 @SuppressWarnings("unused")
 public interface IArdwConfig {
 
-    long BAUD_300 = 300;
-    long BAUD_1200 = 1200;
-    long BAUD_4800 = 4800;
-    long BAUD_9600 = 9600;
-    long BAUD_19200 = 19200;
-    long BAUD_38400 = 38400;
-    long BAUD_57600 = 57600;
-    long BAUD_115200 = 115200;
+    int BAUD_300 = 300;
+    int BAUD_1200 = 1200;
+    int BAUD_4800 = 4800;
+    int BAUD_9600 = 9600;
+    int BAUD_19200 = 19200;
+    int BAUD_38400 = 38400;
+    int BAUD_57600 = 57600;
+    int BAUD_115200 = 115200;
 
 }
